Fix various small problems found with IntelliJ
AbstractServlet:          make constructor protected
Channel:                  unnecessary unboxing
Client:                   make constructor protected
Config:                   use foreach loop
IRCMask:                  make completeWildMask static
ModuleManager:            don't use Class.newInstance
Server:                   make constructor protected, use String.contains over indexOf
commands/Part:            use foreach loop
network/HostResolver:     call super.finalize within finalize method
network/NetworkClient:    use foreach loop, unnecessary throws IOException
network/NetworkServer:    unnecessary casting and unboxing
network/TlsNetworkClient: missing @Override
util/ChannelChecks:       unnecessary initialization
util/ModeUtils:           unnecessary (& 0xFFFFFFFF)

+++ src/main/java/uk/org/cowgill/james/jircd/Channel.java
@@ -922,7 +922,7 @@ public final class Channel
 
 				if(param instanceof Integer)
 				{
-					newLimit = ((Integer) param).intValue();
+					newLimit = (Integer) param;
 				}
 				else
 				{
